+++ packages/canary-docs/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-canary.24](https://github.com/mi6/ic-ui-kit/compare/@ukic/canary-docs@3.0.0-canary.23...@ukic/canary-docs@3.0.0-canary.24) (2025-11-26)
+
+**Note:** Version bump only for package @ukic/canary-docs
+
 # [3.0.0-canary.23](https://github.com/mi6/ic-ui-kit/compare/@ukic/canary-docs@3.0.0-canary.22...@ukic/canary-docs@3.0.0-canary.23) (2025-10-29)
 
 **Note:** Version bump only for package @ukic/canary-docs