+++ packages/docs/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.5](https://github.com/mi6/ic-ui-kit/compare/@ukic/docs@3.4.4...@ukic/docs@3.4.5) (2025-11-26)
+
+**Note:** Version bump only for package @ukic/docs
+
 ## [3.4.4](https://github.com/mi6/ic-ui-kit/compare/@ukic/docs@3.4.3...@ukic/docs@3.4.4) (2025-10-29)
 
 **Note:** Version bump only for package @ukic/docs