+++ packages/nextjs/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.13](https://github.com/mi6/ic-ui-kit/compare/@ukic/nextjs@3.2.12...@ukic/nextjs@3.2.13) (2025-11-26)
+
+**Note:** Version bump only for package @ukic/nextjs
+
 ## [3.2.12](https://github.com/mi6/ic-ui-kit/compare/@ukic/nextjs@3.2.11...@ukic/nextjs@3.2.12) (2025-10-29)
 
 **Note:** Version bump only for package @ukic/nextjs